--- itests/org.openhab.core.persistence.tests/itest.bndrun
@@ -1,120 +1,0 @@
--include: ../itest-common.bndrun
-
-Bundle-SymbolicName: ${project.artifactId}
-Fragment-Host: org.openhab.core.model.persistence
-
--runrequires: \
-	bnd.identity;id='org.openhab.core.persistence.tests',\
-	bnd.identity;id='org.openhab.core.model.persistence.runtime',\
-	bnd.identity;id='org.openhab.core.persistence'
-
-#
-# done
-#
--runbundles: \
-	com.google.inject;version='[3.0.0,3.0.1)',\
-	javax.measure.unit-api;version='[1.0.0,1.0.1)',\
-	log4j;version='[1.2.17,1.2.18)',\
-	org.antlr.runtime;version='[3.2.0,3.2.1)',\
-	org.apache.felix.scr;version='[2.1.10,2.1.11)',\
-	org.eclipse.emf.common;version='[2.12.0,2.12.1)',\
-	org.eclipse.emf.ecore;version='[2.12.0,2.12.1)',\
-	org.eclipse.emf.ecore.xmi;version='[2.12.0,2.12.1)',\
-	org.glassfish.hk2.external.aopalliance-repackaged;version='[2.4.0,2.4.1)',\
-	org.glassfish.hk2.external.javax.inject;version='[2.4.0,2.4.1)',\
-	org.osgi.service.event;version='[1.4.0,1.4.1)',\
-	ch.qos.logback.classic;version='[1.2.3,1.2.4)',\
-	ch.qos.logback.core;version='[1.2.3,1.2.4)',\
-	com.google.gson;version='[2.8.2,2.8.3)',\
-	org.apache.servicemix.specs.activation-api-1.1;version='[2.9.0,2.9.1)',\
-	org.apache.servicemix.specs.annotation-api-1.3;version='[1.3.0,1.3.1)',\
-	org.apache.servicemix.specs.jaxb-api-2.2;version='[2.9.0,2.9.1)',\
-	org.apache.servicemix.specs.stax-api-1.2;version='[2.9.0,2.9.1)',\
-	slf4j.api;version='[1.7.25,1.7.26)',\
-	org.apache.felix.configadmin;version='[1.9.8,1.9.9)',\
-	org.eclipse.equinox.event;version='[1.4.300,1.4.301)',\
-	tec.uom.lib.uom-lib-common;version='[1.0.3,1.0.4)',\
-	tec.uom.se;version='[1.0.10,1.0.11)',\
-	com.google.guava;version='[27.1.0,27.1.1)',\
-	com.google.guava.failureaccess;version='[1.0.1,1.0.2)',\
-	io.github.classgraph;version='[4.8.35,4.8.36)',\
-	org.eclipse.equinox.common;version='[3.10.400,3.10.401)',\
-	org.eclipse.xtend.lib;version='[2.19.0,2.19.1)',\
-	org.eclipse.xtend.lib.macro;version='[2.19.0,2.19.1)',\
-	org.eclipse.xtext;version='[2.19.0,2.19.1)',\
-	org.eclipse.xtext.common.types;version='[2.19.0,2.19.1)',\
-	org.eclipse.xtext.util;version='[2.19.0,2.19.1)',\
-	org.eclipse.xtext.xbase.lib;version='[2.19.0,2.19.1)',\
-	org.objectweb.asm;version='[7.1.0,7.1.1)',\
-	jollyday;version='[0.5.8,0.5.9)',\
-	org.apache.felix.http.servlet-api;version='[1.1.2,1.1.3)',\
-	org.apache.xbean.bundleutils;version='[4.12.0,4.12.1)',\
-	org.apache.xbean.finder;version='[4.12.0,4.12.1)',\
-	org.eclipse.xtext.xbase;version='[2.19.0,2.19.1)',\
-	org.objectweb.asm.commons;version='[7.1.0,7.1.1)',\
-	org.objectweb.asm.tree;version='[7.1.0,7.1.1)',\
-	tec.uom.lib.uom-lib-common;version='[1.0.3,1.0.4)',\
-	tec.uom.se;version='[1.0.10,1.0.11)',\
-	com.google.guava;version='[27.1.0,27.1.1)',\
-	com.google.guava.failureaccess;version='[1.0.1,1.0.2)',\
-	io.github.classgraph;version='[4.8.35,4.8.36)',\
-	org.eclipse.equinox.common;version='[3.10.400,3.10.401)',\
-	org.eclipse.xtend.lib;version='[2.19.0,2.19.1)',\
-	org.eclipse.xtend.lib.macro;version='[2.19.0,2.19.1)',\
-	org.eclipse.xtext;version='[2.19.0,2.19.1)',\
-	org.eclipse.xtext.common.types;version='[2.19.0,2.19.1)',\
-	org.eclipse.xtext.util;version='[2.19.0,2.19.1)',\
-	org.eclipse.xtext.xbase;version='[2.19.0,2.19.1)',\
-	org.eclipse.xtext.xbase.lib;version='[2.19.0,2.19.1)',\
-	org.objectweb.asm;version='[7.1.0,7.1.1)',\
-	org.objectweb.asm.commons;version='[7.1.0,7.1.1)',\
-	org.objectweb.asm.tree;version='[7.1.0,7.1.1)',\
-	org.threeten.extra;version='[1.4.0,1.4.1)',\
-	org.eclipse.jetty.client;version='[9.4.20,9.4.21)',\
-	org.eclipse.jetty.http;version='[9.4.20,9.4.21)',\
-	org.eclipse.jetty.io;version='[9.4.20,9.4.21)',\
-	org.eclipse.jetty.security;version='[9.4.20,9.4.21)',\
-	org.eclipse.jetty.server;version='[9.4.20,9.4.21)',\
-	org.eclipse.jetty.servlet;version='[9.4.20,9.4.21)',\
-	org.eclipse.jetty.util;version='[9.4.20,9.4.21)',\
-	org.eclipse.jetty.websocket.api;version='[9.4.20,9.4.21)',\
-	org.eclipse.jetty.websocket.client;version='[9.4.20,9.4.21)',\
-	org.eclipse.jetty.websocket.common;version='[9.4.20,9.4.21)',\
-	org.eclipse.jetty.xml;version='[9.4.20,9.4.21)',\
-	org.ops4j.pax.swissbox.optional.jcl;version='[1.8.3,1.8.4)',\
-	org.ops4j.pax.web.pax-web-api;version='[7.2.11,7.2.12)',\
-	org.ops4j.pax.web.pax-web-jetty;version='[7.2.11,7.2.12)',\
-	org.ops4j.pax.web.pax-web-runtime;version='[7.2.11,7.2.12)',\
-	org.ops4j.pax.web.pax-web-spi;version='[7.2.11,7.2.12)',\
-	org.openhab.core;version='[3.0.0,3.0.1)',\
-	org.openhab.core.audio;version='[3.0.0,3.0.1)',\
-	org.openhab.core.config.core;version='[3.0.0,3.0.1)',\
-	org.openhab.core.ephemeris;version='[3.0.0,3.0.1)',\
-	org.openhab.core.io.console;version='[3.0.0,3.0.1)',\
-	org.openhab.core.io.http;version='[3.0.0,3.0.1)',\
-	org.openhab.core.io.net;version='[3.0.0,3.0.1)',\
-	org.openhab.core.model.core;version='[3.0.0,3.0.1)',\
-	org.openhab.core.model.item;version='[3.0.0,3.0.1)',\
-	org.openhab.core.model.persistence;version='[3.0.0,3.0.1)',\
-	org.openhab.core.model.persistence.runtime;version='[3.0.0,3.0.1)',\
-	org.openhab.core.persistence.tests;version='[3.0.0,3.0.1)',\
-	org.openhab.core.model.rule;version='[3.0.0,3.0.1)',\
-	org.openhab.core.model.script;version='[3.0.0,3.0.1)',\
-	org.openhab.core.model.script.runtime;version='[3.0.0,3.0.1)',\
-	org.openhab.core.model.sitemap;version='[3.0.0,3.0.1)',\
-	org.openhab.core.model.thing;version='[3.0.0,3.0.1)',\
-	org.openhab.core.persistence;version='[3.0.0,3.0.1)',\
-	org.openhab.core.test;version='[3.0.0,3.0.1)',\
-	org.openhab.core.thing;version='[3.0.0,3.0.1)',\
-	org.openhab.core.transform;version='[3.0.0,3.0.1)',\
-	org.openhab.core.voice;version='[3.0.0,3.0.1)',\
-	org.openhab.core.automation;version='[3.0.0,3.0.1)',\
-	org.openhab.core.automation.module.script;version='[3.0.0,3.0.1)',\
-	biz.aQute.tester.junit-platform;version='[5.1.2,5.1.3)',\
-	junit-jupiter-api;version='[5.6.2,5.6.3)',\
-	junit-jupiter-engine;version='[5.6.2,5.6.3)',\
-	junit-platform-commons;version='[1.6.2,1.6.3)',\
-	junit-platform-engine;version='[1.6.2,1.6.3)',\
-	junit-platform-launcher;version='[1.6.2,1.6.3)',\
-	org.hamcrest;version='[2.2.0,2.2.1)',\
-	org.opentest4j;version='[1.2.0,1.2.1)'
