--- itests/org.openhab.core.persistence.tests/pom.xml
@@ -1,25 +1,0 @@
-<?xml version="1.0" encoding="UTF-8" standalone="no"?>
-<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-  xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
-
-  <modelVersion>4.0.0</modelVersion>
-
-  <parent>
-    <groupId>org.openhab.core.itests</groupId>
-    <artifactId>org.openhab.core.reactor.itests</artifactId>
-    <version>3.0.0-SNAPSHOT</version>
-  </parent>
-
-  <artifactId>org.openhab.core.persistence.tests</artifactId>
-
-  <name>openHAB Core :: Integration Tests :: Persistence Tests</name>
-
-  <dependencies>
-    <dependency>
-      <groupId>org.openhab.core.bom</groupId>
-      <artifactId>org.openhab.core.bom.compile-model</artifactId>
-      <type>pom</type>
-    </dependency>
-  </dependencies>
-
-</project>
